+++ ShopWithMe/Views/Einkaufen/EinkaufenView.swift
@@ -1813,4 +1813,5 @@ private struct MengenNotizSheet: View {
     EinkaufenView()
         .modelContainer(for: [Geschaeft.self, GeschaeftTyp.self, ArtikelKategorie.self, Artikel.self, Einkaufsvorgang.self, Einkaufsliste.self, EinkaufslistenEintrag.self, WarengruppenDistanz.self], inMemory: true)
         .environmentObject(SyncPollingService())
+        .environmentObject(MultipeerSyncService())
 }